requirements.txt
# Cold Email Agent - Python Dependencies

# Core
python-dotenv==1.0.0

# Database
SQLAlchemy==2.0.23

# Email
secure-smtplib==0.1.1

# Web scraping & enrichment
requests==2.31.0
beautifulsoup4==4.12.2

# AI - OpenAI
openai==1.3.0

# Config
PyYAML==6.0.1

# Development & Testing
pytest==7.4.3
pytest-cov==4.1.0
